fix building warning 3

+++ src/lcd800x480/lcd_sueb.c
@@ -379,9 +379,9 @@ u16 LCD_ReadPoint(u16 x,u16 y)
 
 void lcd_clr_window(u16 Color, u16 xs, u16 ys, u16 xe, u16 ye)
 {
-  unsigned int i;//,m;  
+  unsigned int i;//,m;
 	LCD_SetWindows(xs,ys,xe,ye);   
-	for(i=0;i<(xe-xs+1)*(ye-ys+1);i++)
+	for(i=0;i<((unsigned int)xe-xs+1)*(ye-ys+1);i++)
 	{
  //   for(m=0;m<lcddev.width;m++)
   //  {
